Make OMZ optional

+++ setup.sh
@@ -42,13 +42,15 @@ if [[ ! -d "$fzf_dir" ]]; then
     fi
 fi
 
-# Install Oh my ZSH
-ohmyzsh_dir="$HOME/.oh-my-zsh"
-if [[ ! -d "$ohmyzsh_dir" ]]; then
-    export KEEP_ZSHRC=yes
-    if ! sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)" "" --unattended; then
-        echo "WARNING! Oh my ZSH is not installed"
-        rm -rf "$ohmyzsh_dir"
+if [[ "$*" =~ --omz ]]; then
+    # Install Oh my ZSH
+    ohmyzsh_dir="$HOME/.oh-my-zsh"
+    if [[ ! -d "$ohmyzsh_dir" ]]; then
+        export KEEP_ZSHRC=yes
+        if ! sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)" "" --unattended; then
+            echo "WARNING! Oh my ZSH is not installed"
+            rm -rf "$ohmyzsh_dir"
+        fi
     fi
 fi
